refactor: move TagRectangular types to types file
