+++ packages/picasso/src/TagRectangular/index.ts
@@ -1,6 +1,6 @@
 import { OmitInternalProps } from '@toptal/picasso-shared'
 
-import { Props } from './TagRectangular'
+import { Props } from './types'
 
 export { default } from './TagRectangular'
 export type TagRectangularProps = OmitInternalProps<Props>
